pyproject.toml
[tool.poetry]
name = "hometrackertool"
version = "0.1.0"
description = ""
authors = ["Agostinho Ramos <agostinhopina095@gmail.com>"]
maintainers = ["Agostinho Ramos <agostinhopina095@gmail.com>"]
license = "MIT"
homepage = "http://hometrackertool.local/"

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2"
flake8 = "^4.0.1"
black = "^22.3.0"
selenium = "^4.1.3"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-environ = "^0.8.1"
django-cors-headers = "^3.11.0"
channels = "^3.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
